move app translations out of app manager
