--- corehq/apps/app_manager/app_translations/__init__.py
@@ -1,10 +1,0 @@
-from .app_translations import (
-    get_unicode_dicts,
-    process_bulk_app_translation_upload,
-    read_uploaded_app_translation_file,
-    validate_bulk_app_translation_upload,
-    expected_bulk_app_sheet_headers,
-    expected_bulk_app_sheet_rows,
-    update_form_translations,
-    escape_output_value,
-)
