+++ corehq/apps/translations/validator.py
@@ -2,16 +2,16 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 import six
+from django.utils.translation import ugettext_lazy as _
 from memoized import memoized
-from corehq.apps.app_manager.app_translations import (
+from six.moves import zip
+
+from corehq.apps.translations.app_translations import (
     expected_bulk_app_sheet_headers,
     expected_bulk_app_sheet_rows,
     get_unicode_dicts,
 )
 from corehq.apps.translations.const import MODULES_AND_FORMS_SHEET_NAME
-from django.utils.translation import ugettext_lazy as _
-from six.moves import zip
-
 
 COLUMNS_TO_COMPARE = {
     'module_and_form': ['Type', 'sheet_name'],
